[#77]/refactor PathEllipsisMenu 컴포넌트 추출
